+++ drivers/misc/mediatek/video/mt6752/dispsys/ddp_dsi.c
@@ -62,10 +62,10 @@ static int dsi_reg_op_debug;
 		if (cmdq) {\
 			*(unsigned int *)(&r) = ((unsigned int)0x00000000); \
 			r.bit = ~(r.bit);  *(unsigned int *)(&v) = ((unsigned int)0x00000000);\
-			v.bit = value; DISP_REG_MASK(cmdq, &REG, AS_UINT32(&v), AS_UINT32(&r)); \
+			v.bit = value; DISP_REG_MASK(cmdq, (unsigned long)&REG, AS_UINT32(&v), AS_UINT32(&r)); \
 		} else { \
 			mt_reg_sync_writel(INREG32(&REG), &r); r.bit = (value);\
-			DISP_REG_SET(cmdq, &REG, INREG32(&r));\
+			DISP_REG_SET(cmdq, (unsigned long)&REG, INREG32(&r));\
 		}				\
 	} while (0)
